Always look for a streamhandler
Previously when trying run tests locally in lambda services using fleece the root logger would not be setup with a handler. When running in aws it provides a handler. To ensure compatibility between the two, we now always ensure a streamhandler is at least present so error messages don't get swallowed.