connexion: encapsulate `call_api` logic in a FleeceApp class
This is an alternative to the `fleece.connexion.call_api` function which
requires a bunch of parameters to call.
`fleece.connexion.call_api.FleeceApp.call_api`, on the other requires
only an `event` parameter because the other connexion-related parameters
are passed to and stored in the FleeceApp class. The benefit here is
some flexibility in how these utilities are used in code.

For example, if you want to just create a WSGI app and throw events at
it, that was not possible with the previous design of
`fleece.connexion`. Instead, you had to pass all of the required
parameters for creating the app, even if the app has already been
created and cached (which works fine, but it is a bit awkward).

Hopefully these changes can help lead to future improvements in the
interface. I have made several notes in the code about specific things
that could be improved upon, extended, or fixed.