Allow X-Ray support to be disabled via env var
This is a workaround to the lack of X-Ray support in moto at this time.
That is tracked as spulec/moto#1006.

To use, set the environment variable XRAY_DISABLE to any value prior to
running tests locally. Conversely, ensure no such variable is set in
production!

There are no automated tests for this, as there are none for the
monkeypatching used by X-Ray.