validate rax creds after reading config file
Looks like the username and api key were required now as CLI arguments,
as some code that read them from the environment variables had been
moved. Additionally the environment variable RS_API_KEY was changed to
RS_APIKEY, so this PR changes it back. It also changes `rs_apikey_var` to
`rs_api_key_var` for consistency.

This also allows the user to not specify the argument or the environment
variables for the username and api key if they come from different
environment variables set in the config file.