Fix code to work with new ruamel version
Apparently the newer version of ruamel (0.15.66 is what I have) is
spitting back not `dict` but `CommentedMap` instances. This class
has the same duck type as `dict` but since we explicitly check for
`dict` the code stopped working and items quit decrypting and
encrypting. This code fixes it to work with both types of data.