Allow use of default key for ssm parameters
The last change to fleece made it possible to encrypt the rendered
SSM parameters using a custom KMS key, but in the process broke the
ability to use the default key.

That's because boto3's ssm put_parameter call is sort of confusing;
the way to use the default is to avoid passing in the `KeyId` arg
entirely, but the code was setting it to None which led to a runtime
error.